Revert "Merge pull request #64 from BioImageTools/image2"
This reverts commit 83f2c491735440fd78bcb977c332bf346c85f700, reversing
changes made to 01f4dc1ba731ca2f3796eb7de768adec4b78af31.
